01_Areas/Codebases/ailcc/comet_framework/requirements.txt
crewai
crewai[tools]
requests
python-dotenv
